Introduce Assembly Model
The old controller has become the "model", and a new controller has been
added to handle the HTTP requests and delegate to the model. This allows
the model functions to be run in a context other than web requests,
e.g. from other models and controllers, or in unit tests.

The result of this refactor also isolates antibiotic-related
functions into their own model and routes.

N.B. PR to be created after testing on the VPN.